chore: 계층 구조 확립
