+++ src/main/java/in/koreatech/koin/domain/bus/service/CityBusService.java
@@ -12,19 +12,19 @@ import java.util.Set;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import in.koreatech.koin.domain.bus.dto.BusScheduleResponse;
 import in.koreatech.koin.domain.bus.dto.city.CityBusTimetableResponse;
+import in.koreatech.koin.domain.bus.facade.route.CityBusRouteManager;
 import in.koreatech.koin.domain.bus.model.city.CityBusCache;
 import in.koreatech.koin.domain.bus.model.city.CityBusRemainTime;
 import in.koreatech.koin.domain.bus.model.city.CityBusRouteCache;
 import in.koreatech.koin.domain.bus.model.city.CityBusTimetable;
 import in.koreatech.koin.domain.bus.model.enums.BusDirection;
+import in.koreatech.koin.domain.bus.model.enums.BusStation;
 import in.koreatech.koin.domain.bus.model.enums.CityBusDirection;
 import in.koreatech.koin.domain.bus.repository.CityBusCacheRepository;
 import in.koreatech.koin.domain.bus.repository.CityBusRouteCacheRepository;
 import in.koreatech.koin.domain.bus.repository.CityBusTimetableRepository;
-import in.koreatech.koin.domain.bus.dto.BusScheduleResponse;
-import in.koreatech.koin.domain.bus.facade.route.CityBusRouteManager;
-import in.koreatech.koin.domain.bus.model.enums.BusStation;
 import lombok.RequiredArgsConstructor;
 
 @Service
