+++ src/main/java/in/koreatech/koin/domain/bus/service/ShuttleBusService.java
@@ -17,14 +17,14 @@ import org.springframework.transaction.annotation.Transactional;
 import in.koreatech.koin.domain.bus.dto.BusRouteCommand;
 import in.koreatech.koin.domain.bus.dto.BusScheduleResponse;
 import in.koreatech.koin.domain.bus.dto.SingleBusTimeResponse;
-import in.koreatech.koin.domain.bus.model.BusRemainTime;
-import in.koreatech.koin.domain.bus.model.enums.BusType;
-import in.koreatech.koin.domain.bus.facade.route.ShuttleBusRouteManager;
 import in.koreatech.koin.domain.bus.dto.shuttle.BusCourseResponse;
+import in.koreatech.koin.domain.bus.facade.route.ShuttleBusRouteManager;
+import in.koreatech.koin.domain.bus.model.BusRemainTime;
+import in.koreatech.koin.domain.bus.model.enums.BusStation;
+import in.koreatech.koin.domain.bus.model.enums.BusType;
 import in.koreatech.koin.domain.bus.model.shuttle.BusCourse;
 import in.koreatech.koin.domain.bus.model.shuttle.Route;
 import in.koreatech.koin.domain.bus.model.shuttle.SchoolBusTimetable;
-import in.koreatech.koin.domain.bus.model.enums.BusStation;
 import in.koreatech.koin.domain.bus.repository.SchoolBusRepository;
 import lombok.RequiredArgsConstructor;
